Definition text displayed with TextInput widget
The search field results in one or more found objects.
The full_def of it is displayed in a TextInput widget.
(it is used as text output only)
There is not yet a text output widget, apart from the Label widget, but that has no auto scrollbar nor bindow sizing option as the TextInput has.