Aliases TreeView and Aspect TreeView corrected
The aliases and aspects are displayed in a TreeView.
However, only the first value is displayed. It is unclear whether or how additional values on the same row(s) can be added. Therefore, table views might be better suited for the task.
Deletion of earlier content is still to be done.