Tables moved to appearing after selection
The options_table, aliases_table and aspects_table all appeared immediately when the search tab was created. This was changed so that the options table appears only after something is entered in the search field or the uid field. And the aliases table and aspects table only appears after a concept in the options table is selected.