Implement the new TreeTable class from Remi for display of Taxonomy view
The new TreeTable widget, which is a sub class of Remi gui.Table is used instead of MyTable for display of the Taxonomy view.
The mechanism for indentation is developed using hierarchy dict.
Individual things that are classified by kinds in the hierarchy are added.
Display of aspects of the kinds in the tree are corrected.