Improve network view and info view, convert possibilities, individual and kind views
Correct and improve the network view based on Remi TreeTable and documents/info view and display of info in pop up message.
Convert the possibilities view and the Individual (composition) view and the kind view, all from Tkinter to Remi based on SingleRowSelectionTable.
Plus some improvements on comments and some name changes