Major rewrite and Extension with search fields for making new expressions of knowledge
Query_views are now not only used for searching for display, but also for searching for objects that are elements in new expressions of knowledge.
This requires additional search fields widgets that are only created when search_for == 'object', a file button and a function to search for a file with a document.
The major rewrite replaces the various search function by one generalized function. Various areas are commented out for later deletion after thorough testing